# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProConnect API",
        "description": "API de la plateforme ProConnect - Application mobile et portail d'administration",
        "contact": {
            "email": "contact@proconnect.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/admin/stats": {
            "get": {
                "tags": [
                    "Admin - Stats"
                ],
                "summary": "Statistiques globales de la plateforme",
                "operationId": "af16c03fdbae75e0a9c65af3989a6bc8",
                "responses": {
                    "200": {
                        "description": "Statistiques récupérées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "totals": {
                                            "properties": {
                                                "users": {
                                                    "type": "integer",
                                                    "example": 150
                                                },
                                                "companies": {
                                                    "type": "integer",
                                                    "example": 45
                                                },
                                                "posts": {
                                                    "type": "integer",
                                                    "example": 230
                                                },
                                                "admins": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "latest_users": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "latest_posts": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé - Admin uniquement"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/admins": {
            "get": {
                "tags": [
                    "Admin - Users"
                ],
                "summary": "Liste des administrateurs",
                "operationId": "36a0d826401e54a7b9f465e761f4bf08",
                "responses": {
                    "200": {
                        "description": "Liste des administrateurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé - Admin uniquement"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Users"
                ],
                "summary": "Créer un nouvel administrateur",
                "operationId": "4670fc4c6731a232a1fe485045188b13",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "address"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Admin Nouveau"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Rue Admin"
                                    },
                                    "account_type": {
                                        "type": "string",
                                        "enum": [
                                            "private",
                                            "pro"
                                        ],
                                        "example": "pro"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Administrateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "admin": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé - Admin uniquement"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrez le token Sanctum obtenu après connexion",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}